% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.mle.R
\name{ar.mle}
\alias{ar.mle}
\title{ar.mle}
\usage{
ar.mle(
  x,
  aic = TRUE,
  ar.max = NULL,
  get.aic = TRUE,
  order.max = NULL,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  ...
)
}
\arguments{
\item{x}{a univariate or multivariate time series.}

\item{aic}{logical. If TRUE then the Akaike Information Criterion is used to choose the order of the autoregressive model. If FALSE, the model of order order.max is fitted.}

\item{ar.max}{maximum order (or order) of model to fit. Defaults to the smaller of N-1 and 10*log10(N) where N is the number of non-missing observations except for method = "mle" where it is the minimum of this quantity and 12.}

\item{get.aic}{logical. If TRUE the aic of the final model will be returned.}

\item{order.max}{included for backward compatability}

\item{na.action}{function to be called to handle missing values. Currently, via na.action = na.pass, only Yule-Walker method can handle missing values which must be consistent within a time point: either all variables must be missing or none.}

\item{demean}{should a mean be estimated during fitting?}

\item{series}{names for the series. Defaults to deparse (subsititute (x)).}

\item{...}{additional arguments for specific methods.}

\item{method}{character string specifying the method to fit the model. Must be one of the strings in the default argument (the first few characters are sufficient). Defaults to "yule-walker".}

\item{object}{a fit from ar().}

\item{newdata}{data to which to apply the prediction.}

\item{n.ahead}{number of steps ahead at which to predict.}

\item{se.fit}{logical: return estimated standard errors of the prediction error?}
}
\value{
The coefficients of an Autoregressive model, its sigma^2, and possibly the value of its AIC.
}
\description{
Fit an autoregressive time series model to the data using the MLE method, by default selecting the complexity by AIC.
}
\author{
Martyn Plummer. Univariate case of ar.yw, ar.mle and C code for univariate case of ar.burg by B. D. Ripley. Minor changes to R code by Brad Dougherty.
}
