% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.ols.R
\name{ar.ols}
\alias{ar.ols}
\title{ar.ols}
\usage{
ar.ols(
  x,
  aic = TRUE,
  get.aic = get.aic,
  order.max = NULL,
  na.action = na.fail,
  demean = TRUE,
  intercept = demean,
  series = NULL,
  ar.max = NULL,
  ...
)
}
\arguments{
\item{x}{A univariate or multivariate time series.}

\item{aic}{Logical flag. If TRUE then the Akaike Information Criterion is used to choose the order of the autoregressive model. If FALSE, the model of order order.max is fitted.}

\item{order.max}{included for backwards compatibility.}

\item{na.action}{function to be called to handle missing values.}

\item{demean}{should the AR model be for x minus its mean?}

\item{intercept}{should a seperate intercept term be fitted?}

\item{series}{names for teh series. Defulats to deparse (subsitute(x)).}

\item{ar.max}{Maximum order (or order) of model to fit. Defaults to 10*log10(N) where N is the number of observations.}

\item{further}{arguments to be passed to or from methods.}
}
\description{
Fit an autoregressive time series model to the data using the OLS method, by default selecting the complexity by AIC.
}
\author{
Adrian Trapletti, Brian Ripley. Minor revisions by Brad Dougherty.
}
