% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.yw.mts.R
\name{ar.yw.mts}
\alias{ar.yw.mts}
\title{ar.yw.mts}
\usage{
\method{ar.yw}{mts}(
  x,
  aic = TRUE,
  get.aic = TRUE,
  order.max = NULL,
  na.action = na.fail,
  demean = TRUE,
  series = NULL,
  var.method = 1L,
  ar.max = NULL,
  ...
)
}
\arguments{
\item{x}{A univariate or multivariate time series.}

\item{aic}{Logical flag. If TRUE then the Akaike Information Criterion is used to choose the order of the autoregressive model. If FALSE, the model of order order.max is fitted.}

\item{order.max}{included for backwards compatibility.}

\item{na.action}{function to be called to handle missing values.}

\item{demean}{should the AR model be for x minus its mean?}

\item{series}{names for teh series. Defulats to deparse (subsitute(x)).}

\item{ar.max}{Maximum order (or order) of model to fit. Defaults to 10*log10(N) where N is the number of observations.}

\item{intercept}{should a seperate intercept term be fitted?}

\item{further}{arguments to be passed to or from methods.}
}
\description{
Fit an autoregressive time series model to the data using the Yule-Walker method, by default selecting the complexity by AIC.
}
