% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIMA.R
\name{ARIMA}
\alias{ARIMA}
\title{ARIMA}
\usage{
ARIMA(
  x,
  order = c(0, 0, 0),
  seasonal = list(order = c(0, 0, 0), period = NA),
  xreg = NULL,
  include.mean = TRUE,
  transform.pars = TRUE,
  fixed = NULL,
  init = NULL,
  method = c("CSS-ML", "ML", "CSS"),
  n.cond,
  optim.control = list(),
  kappa = 1e+06,
  io = NULL,
  xtransf,
  transfer = NULL,
  include.drift = FALSE,
  include.constant,
  lambda = NULL,
  biasadj = FALSE,
  model = NULL
)
}
\arguments{
\item{order}{regular ARIMA order}

\item{seasonal}{seasonal ARIMA order}

\item{xreg}{a dataframe containing covariates}

\item{transform.pars}{if true, the AR parameters are transformed to ensure stationarity}

\item{fixed}{a vector indicating which coefficients are fixed or free}

\item{init}{initial values}

\item{method}{Fitting method: maximum likelihood or minimize conditional sum-of-squares. The default (unless there are missing values) is to use conditional-sum-of-squares to find starting values, then maximum likelihood.}

\item{n.cond}{number of initial values to be conditioned on in a conditional analysis}

\item{optim.control}{control parameters for the optimization procedure}

\item{kappa}{prior variance; used in dealing with initial values}

\item{io}{a list of time points at which the model may have an innovative outlier. The time point of the outlier can be given either as absolute time point or as c(a,b), i.e. at the b-th 'month' of the a-th 'year' where each year has frequency(x) months, assuming x is a time series.}

\item{xtransf}{xtranf is a matrix with each column containing a covariate that affects the time series response in terms of an ARMA filter of order (p,q), i.e. if Z is one such covariate, its effect on the time series is $(theta_0+theta_1B+...+theta_{q-1}B^{q-1})/(1-phi_1 B -...-phi_p B^p) Z_t$ In particular, if $p=0$ and $q=1$, this specifies a simple regression relationship, which should be included in xreg and not here. Note that the filter starts with zero initial values. Hence, it is pertinent to mean-delete each distributed-lag covariate, and this is not done automatically.}

\item{transfer}{a list consisting of the ARMA orders for each transfer (distributed lag) covariate.}

\item{include.drift}{Should the ARIMA model include a linear drift term? (i.e., a linear regression with ARIMA errors is fitted.) The default is FALSE.}

\item{include.constant}{If TRUE, then include.mean is set to be TRUE for undifferenced series and include.drift is set to be TRUE for differenced series. Note that if there is more than one difference taken, no constant is included regardless of the value of this argument. This is deliberate as otherwise quadratic and higher order polynomial trends would be induced.}

\item{lambda}{Box-Cox transformation parameter. If lambda="auto", then a transformation is automatically selected using BoxCox.lambda. The transformation is ignored if NULL. Otherwise, data transformed before model is estimated.}

\item{biasadj}{Use adjusted back-transformed mean for Box-Cox transformations. If transformed data is used to produce forecasts and fitted values, a regular back transformation will result in median forecasts. If biasadj is TRUE, an adjustment will be made to produce mean forecasts and fitted values.}

\item{model}{Output from a previous call to Arima. If model is passed, this same model is fitted to y without re-estimating any parameters.}

\item{included.mean}{Should the ARIMA model include a mean term? The default is TRUE for undifferenced series, FALSE for differenced ones (where a mean would not affect the fit nor predictions).}

\item{.}{All of the above parameters have the same usage as those in the arima function. Please check the help manual of the arima function. Below are new options.}
}
\description{
This function is very simiar to the arima function in TSA and the arimax function which builds on and extends the capability of the arima function in R stats by allowing the incorporation of transfer functions, and innovative and additive outliers. It differentiates from arima allowing for the calculation of a drift term.
}
\author{

}
