% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMASUBSETS.R
\name{ARMASUBSETS}
\alias{ARMASUBSETS}
\title{ARMASUBSETS}
\usage{
ARMASUBSETS(x, ar.max, ma.max, y.name = "Y", ar.method = "ols", y, ...)
}
\arguments{
\item{x}{time-series data.}

\item{ar.max}{maximum AR order.}

\item{ma.max}{maximum MA order.}

\item{y.name}{label of the time series.}

\item{y}{included for backward compatability.}

\item{...}{arguments passed to the plot.armasubsets function}

\item{method}{method used for fitting the long AR model; default is ols with the AR order determined by AIC.}

\item{nar}{included for backward compatability.}

\item{nma}{included for backward compatability.}
}
\description{
This function finds a number of subset ARMA models. A "long" AR model is fitted to the data y to compute the residuals which are taken as a proxy of the error process. Then, an ARMA model is approximated by a regression model with the the covariates being the lags of the time series and the lags of the error process. Subset ARMA models may then be selected using the subset regression technique by leaps and bounds, via the regsubsets function of the leaps package in R. This function is very similar to the armasubsets function in TSA, with changes having been made to some of the argument names.
}
\author{
This code was orignially implemented by Kung-Sik Chan. Minor changes have been made by Brad Dougherty.
}
