% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOXCOX.AR.R
\name{BOXCOX.AR}
\alias{BOXCOX.AR}
\title{BOXCOX}
\usage{
BOXCOX.AR(
  x,
  order,
  lambda = seq(-2, 2, 0.01),
  plotit = TRUE,
  method = c("mle", "yule-walker", "burg", "ols", "yw"),
  y,
  ...
)
}
\arguments{
\item{x}{univariate time series (must be positive)}

\item{order}{AR order for the data; if missing, the order is determined by AIC for the log-transformed data}

\item{lambda}{a vector of candidate power transformation values; if missing, it is set to be from -2 to 2, with increment .01}

\item{plotit}{logical value, if true, plot the profile log-likelihood for the power estimator}

\item{method}{method of AR estimation; default is "mle"}

\item{y}{Included for backward compatability.}

\item{...}{other parameters to be passed to the ar function}
}
\description{
Determine the appropriate power transformation for time-series data. The objective is to estimate the power transformation so that the transformed time series is approximately a Gaussian AR process. This function is very similar to the armasubsets function in TSA, with changes having been made to some of the argument names.
}
\author{
Kung-Sik Chan, minor revisions were made by Brad Dougherty.
}
