% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA.R
\name{ma}
\alias{ma}
\title{ma()}
\usage{
ma(x, order, centre = TRUE)
}
\arguments{
\item{x}{Univariate time series}

\item{order}{Order of moving average smoother}

\item{centre}{If TRUE, then the moving average is centered for even orders.}
}
\description{
computes a simple moving average smoother of a given time series.
}
\author{
Rob J Hydman.
}
