% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacf.R
\name{pacf}
\alias{pacf}
\title{pacf}
\usage{
pacf(x, lag.max = NULL, plot = TRUE, na.action = na.fail, ...)
}
\arguments{
\item{x}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}

\item{lag.max}{maximum number of lags at which to calculate the acf. Default is 10*log10(N/m) where N is the number of observations and m the number of series.}

\item{plot}{logical. If TRUE (the default) the acf is plotted.}

\item{na.action}{function to be called to handle missing values. na.pass can be used.}

\item{...}{further arguments to be passed to \code{plot.acf}.}

\item{type}{character string giving the type of acf to be computed. Allowed values are "correlation" (the default), "covariance" or "partial".}

\item{demean}{logical. Should the covariances be about the sample means?}

\item{drop.lag.0}{logical. Should lag 0 be dropped.}
}
\value{
An object of class \code{"acf"}, which is a list with the following
  elements:

  \item{lag}{A three dimensional array containing the lags at which
    the acf is estimated.}
  \item{acf}{An array with the same dimensions as \code{lag} containing
    the estimated acf.}
  \item{type}{The type of correlation (same as the \code{type}
    argument).}
  \item{n.used}{The number of observations in the time series.}
  \item{series}{The name of the series \code{x}.}
  \item{snames}{The series names for a multivariate time series.}

  The lag \code{k} value returned by \code{ccf(x, y)} estimates the
  correlation between \code{x[t+k]} and \code{y[t]}.

  The result is returned invisibly if \code{plot} is \code{TRUE}.
}
\description{
Used for the partial autocorrelations.

pacf uses code from the \href{/library/stats/help/pacf}{\code{pacf}} function found in the \href{/library/stats/help/stats-package}{\code{stats}} package.
}
\details{
For \code{type} = \code{"correlation"} and \code{"covariance"}, the
  estimates are based on the sample covariance. (The lag 0 autocorrelation
  is fixed at 1 by convention.)

  By default, no missing values are allowed.  If the \code{na.action}
  function passes through missing values (as \code{na.pass} does), the
  covariances are computed from the complete cases.  This means that the
  estimate computed may well not be a valid autocorrelation sequence,
  and may contain missing values.  Missing values are not allowed when
  computing the PACF of a multivariate time series.

  The partial correlation coefficient is estimated by fitting
  autoregressive models of successively higher orders up to
  \code{lag.max}.

  The generic function \code{plot} has a method for objects of class
  \code{"acf"}.

  The lag is returned and plotted in units of time, and not numbers of
  observations.

  There are \code{print} and subsetting methods for objects of class
  \code{"acf"}.
}
\examples{
require(graphics)

## Examples from Venables & Ripley
acf(lh)
acf(lh, type = "covariance")
pacf(lh)

acf(ldeaths)
acf(ldeaths, ci.type = "ma")
acf(ts.union(mdeaths, fdeaths))
ccf(mdeaths, fdeaths, ylab = "cross-correlation")
# (just the cross-correlations)

presidents # contains missing values
acf(presidents, na.action = na.pass)
pacf(presidents, na.action = na.pass)
}
\references{
Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S}.  Fourth Edition.
  Springer-Verlag.

  (This contains the exact definitions used.)
}
\seealso{
\code{\link[stats]{plot.acf}}, \code{\link[stats]{ARMAacf}} for the exact
  autocorrelations of a given ARMA process.
}
\author{
Original: Paul Gilbert, Martyn Plummer.
  Extensive modifications and univariate case of \code{pacf} by
  B. D. Ripley.
}
