% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.acf.R
\name{plot.acf2}
\alias{plot.acf2}
\title{plot.acf2}
\usage{
\method{plot}{acf2}(
  x,
  xlab = "Lag",
  type = "h",
  ylab = "ACF",
  ci.type = c("white", "ma")[1],
  main = paste("Series", x$series),
  ci.col = "blue",
  ...
)
}
\description{
Plot method for objects of class "acf".
}
