% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ar.R
\name{predict.ar}
\alias{predict.ar}
\title{predict.ar}
\usage{
\method{predict}{ar}(object, newdata, n.ahead = 1L, se.fit = TRUE, ...)
}
\description{
Creates a times series prediction based off of an AR object.
}
